+++ http/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/db/KubernetesClusterComponentSpec.scala
@@ -1,19 +1,14 @@
 package org.broadinstitute.dsde.workbench.leonardo.db
 
-import java.time.Instant
-import org.broadinstitute.dsde.workbench.leonardo.KubernetesTestData._
 import org.broadinstitute.dsde.workbench.leonardo.CommonTestData._
+import org.broadinstitute.dsde.workbench.leonardo.KubernetesTestData._
 import org.broadinstitute.dsde.workbench.leonardo.TestUtils._
-import org.broadinstitute.dsde.workbench.leonardo.{
-  KubernetesClusterAsyncFields,
-  KubernetesClusterStatus,
-  NodepoolStatus
-}
-
-import scala.concurrent.ExecutionContext.Implicits.global
+import org.broadinstitute.dsde.workbench.leonardo.{KubernetesClusterAsyncFields, KubernetesClusterStatus, NodepoolStatus}
 import org.scalatest.flatspec.AnyFlatSpecLike
 
+import java.time.Instant
 import java.time.temporal.ChronoUnit
+import scala.concurrent.ExecutionContext.Implicits.global
 
 class KubernetesClusterComponentSpec extends AnyFlatSpecLike with TestComponent {
 
@@ -31,10 +26,10 @@ class KubernetesClusterComponentSpec extends AnyFlatSpecLike with TestComponent 
     dbFutureValue(kubernetesClusterQuery.getMinimalClusterById(savedCluster2.id)) shouldEqual Some(savedCluster2)
 
     dbFutureValue(
-      kubernetesClusterQuery.getMinimalActiveClusterByName(savedCluster1.cloudContext)
+      kubernetesClusterQuery.getMinimalActiveClusterByCloudContext(savedCluster1.cloudContext)
     ) shouldEqual Some(savedCluster1)
     dbFutureValue(
-      kubernetesClusterQuery.getMinimalActiveClusterByName(savedCluster2.cloudContext)
+      kubernetesClusterQuery.getMinimalActiveClusterByCloudContext(savedCluster2.cloudContext)
     ) shouldEqual Some(savedCluster2)
 
     // should delete the cluster and initial nodepool, hence '2' records updated
