+++ http/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/util/AKSManualTest.scala
@@ -100,8 +100,7 @@ object AKSManualTest {
             DefaultNodepool.fromNodepool(
               cluster.nodepools.headOption
                 .getOrElse(throw new Exception("test clusters to be saved must have at least 1 nodepool"))
-            ),
-            None
+            )
           )
           for {
             saveClusterResult <- dbRef.inTransaction(KubernetesServiceDbQueries.saveOrGetClusterForApp(saveCluster))
